+++ R/simmer-methods.R
@@ -1,6 +1,6 @@
 # Copyright (C) 2014-2015 Bart Smeets
 # Copyright (C) 2015-2016 Bart Smeets and Iñaki Ucar
-# Copyright (C) 2016-2019 Iñaki Ucar
+# Copyright (C) 2016-2021 Iñaki Ucar
 #
 # This file is part of simmer.
 #
@@ -79,7 +79,8 @@ simmer <- function(name="anonymous", verbose=FALSE, mon=monitor_mem(), log_level
 
   env <- list2env(list(
     name=name, mon=mon, resources=list(), sources=list(), globals=list(),
-    sim_obj=Simulator__new(name, verbose, mon$xptr, positive(log_level))))
+    sim_obj=Simulator__new(name, verbose, mon$xptr, positive(log_level))
+  ), parent = asNamespace("simmer"))
 
   class(env) <- "simmer"
   env
@@ -350,7 +351,7 @@ add_resource.simmer <- function(.env, name, capacity=1, queue_size=Inf, mon=TRUE
 #' @param trajectory the trajectory that the generated arrivals will follow (see
 #' \code{\link{trajectory}}).
 #' @param distribution a function modelling the interarrival times (returning a
-#' negative value stops the generator).
+#' negative value or a missing value stops the generator).
 #' @param mon whether the simulator must monitor the generated arrivals or not
 #' (0 = no monitoring, 1 = simple arrival monitoring, 2 = level 1 + arrival
 #' attribute monitoring)
@@ -541,8 +542,10 @@ get_mon_arrivals <- function(.envs, per_resource=FALSE, ongoing=FALSE)
 #' @export
 get_mon_arrivals.simmer <- function(.envs, per_resource=FALSE, ongoing=FALSE) {
   envs_apply(.envs, function(x) {
-    if (ongoing) record_ongoing_(x$sim_obj, per_resource)
-    x$mon$get_arrivals(per_resource)
+    out <- x$mon$get_arrivals(per_resource)
+    if (ongoing)
+      out <- rbind(out, get_ongoing_(x$sim_obj, per_resource))
+    out
   })
 }