+++ R/trajectory-activities.R
@@ -1,6 +1,6 @@
 # Copyright (C) 2014-2015 Bart Smeets
 # Copyright (C) 2015-2016 Bart Smeets and Iñaki Ucar
-# Copyright (C) 2016-2019 Iñaki Ucar
+# Copyright (C) 2016-2019,2021 Iñaki Ucar
 #
 # This file is part of simmer.
 #
@@ -237,7 +237,7 @@ release_selected_all.trajectory <- function(.trj, id=0) {
 #' @inheritParams select
 #' @inheritParams set_attribute
 #' @param resource the name of the resource.
-#' @param value new value to set.
+#' @param value numeric value to set, or a callable object (a function) which must return a numeric value.
 #'
 #' @return Returns the trajectory object.
 #'
@@ -685,8 +685,8 @@ set_trajectory.trajectory <- function(.trj, sources, trajectory) {
 
 #' @rdname set_trajectory
 #' @param object a function modelling the interarrival times (if the source type
-#' is a generator; returning a negative value stops the generator) or a data frame
-#' (if the source type is a data source).
+#' is a generator; returning a negative value or a missing value stops the
+#' generator) or a data frame (if the source type is a data source).
 #' @export
 set_source <- function(.trj, sources, object) UseMethod("set_source")
 
@@ -1139,7 +1139,8 @@ synchronize.trajectory <- function(.trj, wait=TRUE, mon_all=FALSE) {
 #' and splitting a previously established batch.
 #'
 #' @inheritParams seize
-#' @param n batch size, accepts a numeric.
+#' @param n batch size, accepts a numeric or a callable object (a function)
+#' which must return a numeric.
 #' @param timeout set an optional timer which triggers batches every
 #' \code{timeout} time units even if the batch size has not been fulfilled,
 #' accepts a numeric or a callable object (a function) which must return a
@@ -1151,7 +1152,7 @@ synchronize.trajectory <- function(.trj, wait=TRUE, mon_all=FALSE) {
 #' across all your \code{batch} activities.
 #' @param rule an optional callable object (a function) which will be applied to
 #' every arrival to determine whether it should be included into the batch, thus
-#  it must return a boolean.
+#' it must return a boolean.
 #'
 #' @return Returns the trajectory object.
 #'
@@ -1208,15 +1209,19 @@ batch <- function(.trj, n, timeout=0, permanent=FALSE, name="", rule=NULL)
 
 #' @export
 batch.trajectory <- function(.trj, n, timeout=0, permanent=FALSE, name="", rule=NULL) {
-  check_args(n="numeric", timeout=c("numeric", "function"), permanent="flag",
-             name="character", rule=c("function", "NULL"))
+  check_args(n=c("numeric", "function"), timeout=c("numeric", "function"),
+             permanent="flag", name="character", rule=c("function", "NULL"))
 
   switch(
-    binarise(is.function(timeout), is.function(rule)),
+    binarise(is.function(n), is.function(timeout), is.function(rule)),
     add_activity(.trj, Batch__new(positive(n), timeout, permanent, name)),
-    add_activity(.trj, Batch__new_func1(positive(n), timeout, permanent, name)),
-    add_activity(.trj, Batch__new_func2(positive(n), timeout, permanent, name, rule)),
-    add_activity(.trj, Batch__new_func3(positive(n), timeout, permanent, name, rule))
+    add_activity(.trj, Batch__new_func1(n, timeout, permanent, name)),
+    add_activity(.trj, Batch__new_func2(positive(n), timeout, permanent, name)),
+    add_activity(.trj, Batch__new_func3(n, timeout, permanent, name)),
+    add_activity(.trj, Batch__new_func4(positive(n), timeout, permanent, name, rule)),
+    add_activity(.trj, Batch__new_func5(n, timeout, permanent, name, rule)),
+    add_activity(.trj, Batch__new_func6(positive(n), timeout, permanent, name, rule)),
+    add_activity(.trj, Batch__new_func7(n, timeout, permanent, name, rule))
   )
 }