+++ tests/testthat/test-trajectory.R
@@ -1,5 +1,5 @@
 # Copyright (C) 2015-2016 Iñaki Ucar and Bart Smeets
-# Copyright (C) 2016-2018 Iñaki Ucar
+# Copyright (C) 2016-2018,2021 Iñaki Ucar
 #
 # This file is part of simmer.
 #
@@ -15,8 +15,6 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with simmer. If not, see <http://www.gnu.org/licenses/>.
-
-context("basic trajectory functionality")
 
 t0 <- trajectory(verbose = TRUE) %>%
   seize("nurse", 1) %>%
@@ -41,6 +39,8 @@ t0 <- trajectory(verbose = TRUE) %>%
   clone(function() 2, trajectory(verbose = TRUE) %>% timeout(1)) %>%
   synchronize() %>%
   batch(1, rule = function() 0) %>%
+  batch(function() 1, rule = function() 0) %>%
+  batch(function() 1, function() 0, rule = function() 0) %>%
   separate() %>%
   renege_in(function() 1, trajectory(verbose = TRUE) %>% timeout(1)) %>%
   renege_if(function() "1", trajectory(verbose = TRUE) %>% timeout(1)) %>%
@@ -84,6 +84,8 @@ trajs <- c(
                                        trajectory(verbose = TRUE) %>% timeout(1)),
   trajectory(verbose = TRUE) %>% synchronize(),
   trajectory(verbose = TRUE) %>% batch(1, rule = function() 0),
+  trajectory(verbose = TRUE) %>% batch(function() 1, rule = function() 0),
+  trajectory(verbose = TRUE) %>% batch(function() 1, function() 0, rule = function() 0),
   trajectory(verbose = TRUE) %>% separate(),
   trajectory(verbose = TRUE) %>% renege_in(function() 1,
                                            trajectory(verbose = TRUE) %>% timeout(1)),